fix: changed product price type

+++ refactoring/src/main/java/ru/akirakozov/sd/refactoring/entities/Product.java
@@ -2,9 +2,9 @@ package ru.akirakozov.sd.refactoring.entities;
 
 public class Product {
     private final String name;
-    private final int price;
+    private final long price;
 
-    public Product(String name, int price) {
+    public Product(String name, long price) {
         this.name = name;
         this.price = price;
     }
@@ -13,7 +13,7 @@ public class Product {
         return name;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 }
